Use trait instead of parent class so it can be sideloaded into other subclasses of TestCase
